pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coding-agent-kit"
version = "0.1.0"
description = "Unified CLI to install and run coding agents for research"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["PyYAML>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=9.0.2", "flake8>=7.1.1"]

[project.scripts]
cakit = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
